requirements.txt
python-dotenv==1.1.1
ollama==0.5.3
transformers==4.55.4
